refactor nats client out of gateway
